Add validation of results
Improve parsing of input for validation datasets

WIP: verify results with validation datasets

Fix ReadData precision + results validation

Match ComputeInvariantMass with ROOT invmass func

Change pt,eta,phi,e from float to double

Fix wrong results

+++ datastructures.h
@@ -4,17 +4,17 @@
 #include "struct_transformer.h"
 
 struct Particle {
-    float pt;
-    float eta;
-    float phi;
-    float e;
+    double pt;
+    double eta;
+    double phi;
+    double e;
 };
 
 struct ParticleRef {
-    float& pt;
-    float& eta;
-    float& phi;
-    float& e;
+    double& pt;
+    double& eta;
+    double& phi;
+    double& e;
 };
 
 template <auto Members> struct SubParticle;
